Maps that got fixed and need wiping (#339)
* Maps that got fixed and need wiping

MY FORMATTING IS ENTIRELY INCORRECT AND I HAVE NO IDEA HOW TO DO THIS CORRECTLY OR IF THIS IS IN THE RIGHT PLACE quat pls help me these fixes were like alreayd put in the game (See #ingame-ticket-468 in the Discord Server for more info) and these times needed deletion. Hopefully all the necessary information should be there to set this up correctly.

* Add Description column

* remove dashes

* write descriptions

* add date of last mapfix

* stupid comma

---------

Co-authored-by: Quaternions <krakow20@gmail.com>